+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.18 branch
 ### 2.18.2
 * updated for MeTools 2.12.5;
+* updated for cakephp-thumber 1.1.0;
 * updated for cakephp-mysql-backup 1.0.3.
 
 ### 2.18.1